backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
umap-learn
scikit-learn
numpy
pydantic
groq
python-dotenv
